+++ lib/reform/form/active_model.rb
@@ -1,13 +1,11 @@
-require "uber/delegates"
-
 module Reform::Form::ActiveModel
   def self.included(base)
     base.class_eval do
       extend ClassMethods
       register_feature ActiveModel
 
-      extend Uber::Delegates
-      delegates :model, *[:persisted?, :to_key, :to_param, :id] # Uber::Delegates
+      extend Forwardable
+      def_delegators :model, :persisted?, :to_key, :to_param, :id
 
       def to_model # this is called somewhere in FormBuilder and ActionController.
         self
@@ -42,15 +40,19 @@ module Reform::Form::ActiveModel
 
     # moved from reform as not applicable to dry
     def validates(*args, &block)
-      validation(:default, inherit: true) { validates *args, &block }
+      validation(name: :default, inherit: true) { validates *args, &block }
     end
 
     def validate(*args, &block)
-      validation(:default, inherit: true) { validate *args, &block }
+      validation(name: :default, inherit: true) { validate *args, &block }
     end
 
     def validates_with(*args, &block)
-      validation(:default, inherit: true) { validates_with *args, &block }
+      validation(name: :default, inherit: true) { validates_with *args, &block }
+    end
+
+    def validates_each(*args, &block)
+      validation(name: :default, inherit: true) { validates_each *args, &block }
     end
 
     # Set a model name for this form if the infered is wrong.
@@ -85,7 +87,6 @@ module Reform::Form::ActiveModel
 
   private
     def active_model_name_for(string, namespace=nil)
-      return ::ActiveModel::Name.new(OpenStruct.new(:name => string)) if ::ActiveModel::VERSION::MAJOR == 3 and ::ActiveModel::VERSION::MINOR == 0 # TODO: remove when we drop rails 3.x.
       ::ActiveModel::Name.new(self, namespace, string)
     end
   end # ClassMethods