+++ lib/reform/form/validation/unique_validator.rb
@@ -7,6 +7,12 @@
 # validates :title, unique: true
 #
 # == Options
+#
+# = Case Sensitivity
+# Case sensitivity is true by default, but can be set to false:
+#
+# validates :title, unique: { case_sensitive: false }
+#
 # = Scope
 # A scope can be use to filter the records that need to be compare with the
 # current value to validate. A scope array can have one to many fields define.
@@ -31,9 +37,19 @@
 class Reform::Form::UniqueValidator < ActiveModel::EachValidator
   def validate_each(form, attribute, value)
     model = form.model_for_property(attribute)
+    original_attribute = form.options_for(attribute)[:private_name]
 
     # search for models with attribute equals to form field value
-    query = model.class.where(attribute => value)
+    query = if options[:case_sensitive] == false && value
+              model.class.where("lower(#{original_attribute}) = ?", value.downcase)
+            else
+              model.class.where(original_attribute => value)
+            end
+
+    # if model persisted, query should bypass model
+    if model.persisted?
+      query = query.where("#{model.class.primary_key} != ?", model.id)
+    end
 
     # apply scope if options has been declared
     Array(options[:scope]).each do |field|
@@ -41,9 +57,7 @@ class Reform::Form::UniqueValidator < ActiveModel::EachValidator
       query = query.where(field => form.send(field))
     end
 
-    # if model persisted, query may return 0 or 1 rows, else 0
-    allow_count = model.persisted? ? 1 : 0
-    form.errors.add(attribute, :taken) if query.count > allow_count
+    form.errors.add(attribute, :taken) if query.count > 0
   end
 end